999.021-libnfs-1.3_support.patch

+++ xbmc/filesystem/FileNFS.cpp
@@ -365,7 +365,7 @@ void CNfsConnection::resetKeepAlive(struct nfsfh  *_pFileHandle)
 //we were before
 void CNfsConnection::keepAlive(struct nfsfh  *_pFileHandle)
 {
-  off64_t offset = 0;
+  uint64_t offset = 0;
   char buffer[32];
   CLog::Log(LOGNOTICE, "NFS: sending keep alive after %i s.",KEEP_ALIVE_TIMEOUT/2);
   CSingleLock lock(*this);
@@ -451,7 +451,7 @@ CFileNFS::~CFileNFS()
 int64_t CFileNFS::GetPosition()
 {
   int ret = 0;
-  off64_t offset = 0;
+  uint64_t offset = 0;
   CSingleLock lock(gNfsConnection);
   
   if (gNfsConnection.GetNfsContext() == NULL || m_pFileHandle == NULL) return 0;
@@ -577,7 +577,7 @@ unsigned int CFileNFS::Read(void *lpBuf, int64_t uiBufSize)
   
   if (m_pFileHandle == NULL || m_pNfsContext == NULL ) return 0;
 
-  numberOfBytesRead = gNfsConnection.GetImpl()->nfs_read(m_pNfsContext, m_pFileHandle, (size_t)uiBufSize, (char *)lpBuf);  
+  numberOfBytesRead = gNfsConnection.GetImpl()->nfs_read(m_pNfsContext, m_pFileHandle, uiBufSize, (char *)lpBuf);  
 
   lock.Leave();//no need to keep the connection lock after that
   
@@ -595,7 +595,7 @@ unsigned int CFileNFS::Read(void *lpBuf, int64_t uiBufSize)
 int64_t CFileNFS::Seek(int64_t iFilePosition, int iWhence)
 {
   int ret = 0;
-  off64_t offset = 0;
+  uint64_t offset = 0;
 
   CSingleLock lock(gNfsConnection);  
   if (m_pFileHandle == NULL || m_pNfsContext == NULL) return -1;
@@ -657,7 +657,7 @@ int CFileNFS::Write(const void* lpBuf, int64_t uiBufSize)
     //write chunk
     writtenBytes = gNfsConnection.GetImpl()->nfs_write(m_pNfsContext,
                                   m_pFileHandle, 
-                                  (size_t)chunkSize, 
+                                  chunkSize, 
                                   (char *)lpBuf + numberOfBytesWritten);
     //decrease left bytes
     leftBytes-= writtenBytes;
